Burnt Rubble & Grass
When we burn buildings using flamethrower, we now replace the building with a burnt rubble. It was simply disappearing before... still want to add more i.e. burning particle effect but it's at least something for the momenet.

Added some grass prefabs and a few rocks., just to add a little more to the map