Cleanup 3 & House Adjustments
Removed more assets that we are not using.

The majority of our memory is coming from the rubble pack's textures in Environment. I've cut it down about as much as I can, sitting at 2.43GB (including 696 MB .git folder).
Not too big to worry right now, but if we are adding more it might be best to try and fix the texture sizes.

* Cleaned up a lot of prefabs in the pack, and removed lots of colliders that are not needed
* If we throw enemies or cars at houses, they will be destroyed now.
* Made colliders on houses a bit smaller to avoid destroying multiple at once due to big colliders.