Minimap
Minimap added.

Enemies show up as red on the minimap and the player is white.

Looks nice, and it also makes it easy to tell when there is an enemy there and they are quite difficult to see. Considered adding health bars to the enemies but it doesn't make much sense to as they should explode after being thrown into buildings, figured this was a better alternative.