Bug Fix
Fixed a bug with traffic light system not deactivating properly.

Trying to check if a gameobject's component was null was throwing a null reference exception before the check was carried out because it was null. Had to create an extension method to check, and change the order of execution in the trigger.

Cars were blowing up when we drop them because it was hitting the players collider, I changed them to use triggers instead of collisions to stop this.

Made the player's sphere collider a bit smaller for more accuracy as he was able to knock over traffic lights before he visually hit them.