Metal Chain Link Fence now uses ForceObject set-up
Removed ability to pick them up as it doesn't seem to add much to the game, and seemed to be a slight collider issue when picking up multiple up at once.

Uses the forceObject set-up so they fall over more naturally and it's more satisfying. Before they all fell exactly the same and looked odd.